+++ src/features/scoped-paths/scoped-paths-workspace-state.ts
@@ -8,7 +8,7 @@ export class ScopedPathsWorkspaceState {
 
   private _enabled: boolean = defaultEnabled
   private _currentScope: string = defaultCurrentScope
-  private _temporaryUnscopedPathsSnapshot: string[] | undefined = undefined
+  private _quickUnscopePathsSnapshot: string[] | undefined = undefined
 
   constructor(
     private readonly workspaceState: Memento,
@@ -19,13 +19,13 @@ export class ScopedPathsWorkspaceState {
   private load() {
     const enabled = this.workspaceState.get<boolean>('streamline.scopedPaths.enabled', defaultEnabled)
     const currentScope = this.workspaceState.get<string>('streamline.scopedPaths.currentScope', defaultCurrentScope)
-    const temporaryUnscopedPathsSnapshot = this.workspaceState.get<string[] | undefined>('streamline.scopedPaths.temporaryUnscopedPathsSnapshot', undefined)
+    const quickUnscopedPathsSnapshot = this.workspaceState.get<string[] | undefined>('streamline.scopedPaths.quickUnscopedPathsSnapshot', undefined)
 
     this._enabled = enabled
     this._currentScope = currentScope
-    this._temporaryUnscopedPathsSnapshot = temporaryUnscopedPathsSnapshot
+    this._quickUnscopePathsSnapshot = quickUnscopedPathsSnapshot
 
-    console.debug('[ScopedPaths] WorkspaceState has been loaded', { enabled, currentScope, temporaryUnscopedPathsSnapshot })
+    console.debug('[ScopedPaths] WorkspaceState has been loaded', { enabled, currentScope, quickUnscopedPathsSnapshot })
   }
 
   async save() {
@@ -40,8 +40,8 @@ export class ScopedPathsWorkspaceState {
     )
 
     await this.workspaceState.update(
-      'streamline.scopedPaths.temporaryUnscopedPathsSnapshot',
-      this._temporaryUnscopedPathsSnapshot,
+      'streamline.scopedPaths.quickUnscopedPathsSnapshot',
+      this._quickUnscopePathsSnapshot,
     )
 
     console.debug('[ScopedPaths] WorkspaceState has been saved')
@@ -65,12 +65,16 @@ export class ScopedPathsWorkspaceState {
     return this._currentScope
   }
 
-  setTemporaryUnscopedPathsSnapshot(value: string[] | undefined) {
-    this._temporaryUnscopedPathsSnapshot = value
+  setQuickUnscopePathsSnapshot(value: string[] | undefined) {
+    this._quickUnscopePathsSnapshot = value
     this.onChange?.()
   }
 
-  getTemporaryUnscopedPathsSnapshot() {
-    return this._temporaryUnscopedPathsSnapshot
+  getQuickUnscopePathsSnapshot() {
+    return this._quickUnscopePathsSnapshot
+  }
+
+  getQuickUnscopeEnabled() {
+    return this._quickUnscopePathsSnapshot !== undefined
   }
 }